RF: Move all version play from info to pkg_info
